requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
Django==4.2.4
idna==3.4
lxml==4.9.3
msal==1.24.0
mssql-django==1.3
Office365-REST-Python-Client==2.4.4
pycparser==2.21
PyJWT==2.8.0
pyodbc==4.0.39
pyspnego==0.9.2
pytz==2023.3
requests==2.31.0
requests-ntlm==1.2.0
requests-toolbelt==1.0.0
SharePlum==0.5.1
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.4
